+++ examples/my_2_outputs_script.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+# This scripts just for use by example myscript_has_two_outputs.py
+entree=$1
+sortie=$2
+autre_sortie=$3
+cp $entree $sortie
+cp $entree $autre_sortie